Add crawl /log API endpoint
If a crawl is completed, the endpoint streams the logs from the log
files in all of the created WACZ files, sorted by timestamp.

If a crawl is still running, the endpoint streams the logs from the
crawler pods live, without sorting by timestamp.

The API endpoint supports filtering by log_level and context whether
the crawl is still running or not.