Store search data directly in Crawl and CrawlConfig collections
In crawls and workflows, instead of computing name and firstSeed
in the aggregation, store these fields directly in the collections.

Additionally, store search-formatted versions of these strings that
strip http(s):// prefixes and trailing slashes to help with search.

This is a step to backend fuzzy search of these fields.