Change checkbox labels to use sentence case
I know we originally said we would change all the form labels to use this, but I didn't like how it looked.  They are all titles and if we're keeping title case for titles that's probably the way to go?

Checkboxes titles are probably _also_ titles... But they look better like this so sentence case it is!