Modify QA state sort
Rename activeQAState sort option to just qaState, which sorts by
activeQAState first, then lastQAState. Both fields are returned
with the data now.

This requires a fair bit of computation in the Mongo aggregation
so we may want to consider moving to precomputing like with workflow
last crawl information at some point. That would also allow the sort
to be more visible based on a single field.