Rename qa sorts to lastQAStarted, lastQAState
Include active QA in these (will already become first since we're
sorted by started) instead of separating it out into a separate
field. This is more consistent with how we handle workflows.

We do however keep activeQAStats, as that can be used to quickly
determine if there is an active QA run and show a progress bar for
it if so.