Always use oid prefix in s3 storage
Previously, files in a default bucket were prefixed with the oid
but were not in custom storages. This commit removes that
distinction to aid in copying files in buckets, removing the need
for unnecessary filepath manipulation.

The CopyBucketJob now only copies an organization's directory
rather than the entire bucket to prevent accidentally copying
another organization's data.