Filter (re-)add all pages endpoint by crawl type in path
If a user POSTS to /crawls/all/ or /uploads/all/, filter the
crawl objects in the database by that type before adding pages.
This enables us to easily add pages to the database for all
uploads in an organization, for example.

GET endpoints and endpoints that specify just a single crawl have
been kept more permissive as long as the crawl_id is found.