Ensure PR workflows can run on all PRs (but still skip when they're not needed) (#2237)
Closes #2279

This adds a `paths-filter` step to all workflows that run on PRs so that
we can enable auto-merge, for more info about this read [enabling
auto-merge](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/incorporating-changes-from-a-pull-request/automatically-merging-a-pull-request)!
In order to be able to have required checks, a workflow can't be
entirely skipped: see [Handling skipped but required
checks](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/collaborating-on-repositories-with-code-quality-features/troubleshooting-required-status-checks#handling-skipped-but-required-checks).

This also merges frontend CI workflows into one with multiple parallel
steps, which should speed things up a bit. It also upgrades node
versions to 20 and 22 across the board.