Hacky workaroud: Pass default org to retry_background_job
Now that we have background jobs that aren't tied to a specific org,
much of the background_jobs module needs to be reworked to account
for that.

That will take some time, so for now, so that we can test the
migration, we just pass the default org to retry_background_job
if the job doesn't have an oid set.