Move crawl and QA logs to new mongo collection (#2791)
Fixes #2765 

(This is a necessary part of ensuring mongo documents don't exceed 16MB.
Hopefully it's also sufficient but we'll need to see in practice if
there are other fields that need to be separated out.)

This PR moves crawl and QA run logs into a separate `crawl_logs` mongo
collection.

It adds a new backend module (without a distinct API router, but the
crawls module is getting quite large and it seemed to make sense to add
a separate module for the new mongo collection), as well as a migration
to move crawl logs from `Crawl` objects into the new collection. The
existing nightly test for crawl error logs is fleshed out, and a new
nightly test added for behavior logs.

---------
Co-authored-by: Ilya Kreymer <ikreymer@users.noreply.github.com>
Co-authored-by: Ilya Kreymer <ikreymer@gmail.com>