Download archived items as single WACZ file when possible via new download endpoint query parameter (#2850)
Fixes #2648 

Replaces https://github.com/webrecorder/browsertrix/pull/2805

This PR introduces a `preferSingleWACZ` query parameter to the
`/all-crawls/<crawl_id>/download` and `/crawls/<crawl_id>/download`
endpoints. When set to true, these endpoints will only create
multi-WACZs when a crawl has more than one WACZ file, and otherwise will
stream the original crawl WACZ.

This flag is not enabled by default to prevent introducing breaking
changes to the API, but the frontend is updated to use it in all places
where it seemed appropriate.

A new backend test is also added to account for the change.
---------

Co-authored-by: Ilya Kreymer <ikreymer@gmail.com>