Fix browser profile launch by specifying relative path inside oid dir to profilejob (#2912)
Fixes #2911 

Follow-up to https://github.com/webrecorder/browsertrix/issues/2891

This commit strips the oid prefix from the profile filename before
passing it along to the profilejob so that it's in the format the
crawler expects.

I've also added a profile to an existing test that runs a crawl and
verifies that state is `complete` at finished (i.e. not failed) to catch
any similar type regression in the future.