WIP: Don't double-count already-uploaded WACZs in check
Needs to be tested, just pushing as-is so that I can pick it up
next week. There's an issue in local testing where crawls sometimes
appear to be twice as big as they really are, which is making
Browsertrix think the storage quota is reached prematurely. I
haven't yet pinned down the cause of this and it seems intermittent.