Use dedicated mongo session in quota update method (#3057)
To be merged into #2944 

Adds a causally consistent session specifically to the quota update
method so that its three operations can operate more-or-less atomically.

We'd looked at doing this with a transaction, but they require a replica
set or sharding, which isn't always included in self-hosted Browsertrix
deployments. For now this should improve reliability for the
multi-operation `update_quotas` method, even if it's not perfectly
atomic — it's pretty unlikely at the scale we're working with here for
there to be multiple simultaneous operations for a single org.

Tested with local deployment.

---------

Co-authored-by: Ilya Kreymer <ikreymer@gmail.com>