Add migration to fix failed bg jobs in database
Some background jobs previously failed and did not have success
or finished fields set due to bugs. This migration targets these
jobs and ensures that their finished and success values are set
in the database as expected.

Only jobs that are at least (replica deletion delay + 1) or 7 days
old are targeted to prevent modifying any jobs that are still in
progress or subject to the replica deletion delay.