Add migrations to fix failed bg jobs, replicate unreplicated files
Some background jobs previously failed and did not have success
or finished fields set due to bugs. The first migration targets
these and ensures that their finished and success values are set
in the database as expected.

The second migration looks for all unreplicated crawl and profile
files and replicates them. This is preferably to simply
retrying older failed replication background jobs, as it is
possible that the objects they correlate have been deleted or
changed since and so those old jobs are no longer applicable.